+++ client/src/components/ModulesSection/ModulesInfoSeller.js
@@ -25,15 +25,15 @@ export const ModulesInfoSeller = [
     modulesItems: [
       {
         title: "Adicionar",
-        url: "/products",
+        url: "/productsAdd",
       },
       {
         title: "Modificar",
-        url: "/products",
+        url: "/productsModify",
       },
       {
         title: "Consultar",
-        url: "/products",
+        url: "/productsModify",
       },
     ],
   },
