+++ client/src/components/UsersManagementSection/UsersManagementSection.js
@@ -4,10 +4,11 @@ import { CreateUserItems } from "./CreateUserItems";
 import { UpdateUserItems } from "./UpdateUserItems";
 import { InputUserItems } from "./InputUserItems";
 import UserContext from "../../context/UserContext";
-import registerUser from "../../assets/register_user.svg";
-import visualizeUser from "../../assets/visualize_user.svg";
 import Auth0Hook from "../../hooks/Auth0Hook";
 import UserManagementTable from "../UserManagementTable/UserManagementTable";
+import registerUserIcon from "../../assets/register_user.svg";
+import updateUserIcon from "../../assets/update_user.svg";
+import visualizeUserIcon from "../../assets/visualize_user.svg";
 
 const UsersManagementSection = () => {
   const { searchUser, addUser, updateUser } = UserContext();
@@ -37,26 +38,8 @@ const UsersManagementSection = () => {
   return (
     <div className="users__management__section">
       <div className="users__section_boxes">
-        <ManagementBox
-          img={registerUser}
-          title="Registro de usuarios"
-          onSubmitFunct={onSubmitRegister}
-          obj={user}
-          formId="CreateUser"
-          buttonName="Añadir usuario"
-          itemsInput={InputUserItems}
-          itemsSelect={CreateUserItems}
-        />
-        <ManagementBox
-          img={visualizeUser}
-          title="Modificar información"
-          onSubmitFunct={onSubmitUpdate}
-          obj={user}
-          formId="UpdateUser"
-          buttonName="Modificar"
-          itemsInput={InputUserItems}
-          itemsSelect={UpdateUserItems}
-        />
+        <ManagementBox  img={registerUserIcon} onSubmitFunct = {onSubmitRegister} obj={user} formId= "CreateUser"  name="Registro de usuarios" buttonName = "Añadir usuario"   itemsInput={InputUserItems} itemsSelect={CreateUserItems} />
+        <ManagementBox img={updateUserIcon}  onSubmitFunct = {onSubmitUpdate}   obj={user}  formId= "UpdateUser"    name="Modificar información" buttonName = "Modificar"  itemsInput={InputUserItems}   itemsSelect={UpdateUserItems}  />
         <UserManagementTable></UserManagementTable>
       </div>
     </div>
